# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Blog API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  /** @nullable */
  coverImageUrl: string | null;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  countryId: number | null;
  /** @nullable */
  publishedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePostBody {
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  countryId?: number | null;
  /** @nullable */
  publishedAt?: string | null;
}

export interface UpdatePostBody {
  title?: string;
  slug?: string;
  content?: string;
  excerpt?: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  countryId?: number | null;
  /** @nullable */
  publishedAt?: string | null;
}

export interface Country {
  id: number;
  name: string;
  countryCode: string;
  visitedCities: string;
  reasonForVisit: string;
  travelCompanions: string;
  friendsFamilyMet: string;
  visitedAt: string;
  latitude: number;
  longitude: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCountryBody {
  name: string;
  countryCode: string;
  visitedCities: string;
  reasonForVisit: string;
  travelCompanions: string;
  friendsFamilyMet: string;
  visitedAt: string;
  latitude: number;
  longitude: number;
}

export interface UpdateCountryBody {
  name?: string;
  countryCode?: string;
  visitedCities?: string;
  reasonForVisit?: string;
  travelCompanions?: string;
  friendsFamilyMet?: string;
  visitedAt?: string;
  latitude?: number;
  longitude?: number;
}

export interface MapPin {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  /** @nullable */
  coverImageUrl: string | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  location: string | null;
  /** @nullable */
  publishedAt: string | null;
}

export interface TravelStats {
  totalCountries: number;
  totalPosts: number;
  totalCities: number;
  continents: number;
}