# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Travel Blog API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Blog post operations
  - name: countries
    description: Country visit operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all blog posts
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a blog post by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /countries:
    get:
      operationId: listCountries
      tags: [countries]
      summary: List all visited countries
      responses:
        "200":
          description: List of visited countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"
    post:
      operationId: createCountry
      tags: [countries]
      summary: Add a visited country
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCountryBody"
      responses:
        "201":
          description: Created country
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"

  /countries/{id}:
    get:
      operationId: getCountry
      tags: [countries]
      summary: Get a visited country by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A visited country
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
        "404":
          description: Not found
    patch:
      operationId: updateCountry
      tags: [countries]
      summary: Update a visited country
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCountryBody"
      responses:
        "200":
          description: Updated country
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
        "404":
          description: Not found
    delete:
      operationId: deleteCountry
      tags: [countries]
      summary: Delete a visited country
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /map-pins:
    get:
      operationId: listMapPins
      tags: [posts]
      summary: Get all map pins (posts with lat/lng) for the world map
      responses:
        "200":
          description: List of map pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapPin"

  /stats:
    get:
      operationId: getStats
      tags: [countries]
      summary: Get travel stats summary
      responses:
        "200":
          description: Travel stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
        countryId:
          type: ["integer", "null"]
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - slug
        - content
        - excerpt
        - coverImageUrl
        - latitude
        - longitude
        - location
        - countryId
        - publishedAt
        - createdAt
        - updatedAt

    CreatePostBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
        countryId:
          type: ["integer", "null"]
        publishedAt:
          type: ["string", "null"]
      required:
        - title
        - slug
        - content
        - excerpt

    UpdatePostBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
        countryId:
          type: ["integer", "null"]
        publishedAt:
          type: ["string", "null"]

    Country:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        countryCode:
          type: string
        visitedCities:
          type: string
        reasonForVisit:
          type: string
        travelCompanions:
          type: string
        friendsFamilyMet:
          type: string
        visitedAt:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - countryCode
        - visitedCities
        - reasonForVisit
        - travelCompanions
        - friendsFamilyMet
        - visitedAt
        - latitude
        - longitude
        - createdAt
        - updatedAt

    CreateCountryBody:
      type: object
      properties:
        name:
          type: string
        countryCode:
          type: string
        visitedCities:
          type: string
        reasonForVisit:
          type: string
        travelCompanions:
          type: string
        friendsFamilyMet:
          type: string
        visitedAt:
          type: string
        latitude:
          type: number
        longitude:
          type: number
      required:
        - name
        - countryCode
        - visitedCities
        - reasonForVisit
        - travelCompanions
        - friendsFamilyMet
        - visitedAt
        - latitude
        - longitude

    UpdateCountryBody:
      type: object
      properties:
        name:
          type: string
        countryCode:
          type: string
        visitedCities:
          type: string
        reasonForVisit:
          type: string
        travelCompanions:
          type: string
        friendsFamilyMet:
          type: string
        visitedAt:
          type: string
        latitude:
          type: number
        longitude:
          type: number

    MapPin:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        latitude:
          type: number
        longitude:
          type: number
        location:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
      required:
        - id
        - title
        - slug
        - excerpt
        - coverImageUrl
        - latitude
        - longitude
        - location
        - publishedAt

    TravelStats:
      type: object
      properties:
        totalCountries:
          type: integer
        totalPosts:
          type: integer
        totalCities:
          type: integer
        continents:
          type: integer
      required:
        - totalCountries
        - totalPosts
        - totalCities
        - continents