# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Blog API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all blog posts
 */
export const ListPostsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  content: zod.string(),
  excerpt: zod.string(),
  coverImageUrl: zod.string().nullable(),
  latitude: zod.number().nullable(),
  longitude: zod.number().nullable(),
  location: zod.string().nullable(),
  countryId: zod.number().nullable(),
  publishedAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListPostsResponse = zod.array(ListPostsResponseItem);

/**
 * @summary Create a blog post
 */
export const CreatePostBody = zod.object({
  title: zod.string(),
  slug: zod.string(),
  content: zod.string(),
  excerpt: zod.string(),
  coverImageUrl: zod.string().nullish(),
  latitude: zod.number().nullish(),
  longitude: zod.number().nullish(),
  location: zod.string().nullish(),
  countryId: zod.number().nullish(),
  publishedAt: zod.string().nullish(),
});

/**
 * @summary Get a blog post by id
 */
export const GetPostParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  content: zod.string(),
  excerpt: zod.string(),
  coverImageUrl: zod.string().nullable(),
  latitude: zod.number().nullable(),
  longitude: zod.number().nullable(),
  location: zod.string().nullable(),
  countryId: zod.number().nullable(),
  publishedAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a blog post
 */
export const UpdatePostParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePostBody = zod.object({
  title: zod.string().optional(),
  slug: zod.string().optional(),
  content: zod.string().optional(),
  excerpt: zod.string().optional(),
  coverImageUrl: zod.string().nullish(),
  latitude: zod.number().nullish(),
  longitude: zod.number().nullish(),
  location: zod.string().nullish(),
  countryId: zod.number().nullish(),
  publishedAt: zod.string().nullish(),
});

export const UpdatePostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  content: zod.string(),
  excerpt: zod.string(),
  coverImageUrl: zod.string().nullable(),
  latitude: zod.number().nullable(),
  longitude: zod.number().nullable(),
  location: zod.string().nullable(),
  countryId: zod.number().nullable(),
  publishedAt: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a blog post
 */
export const DeletePostParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all visited countries
 */
export const ListCountriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  countryCode: zod.string(),
  visitedCities: zod.string(),
  reasonForVisit: zod.string(),
  travelCompanions: zod.string(),
  friendsFamilyMet: zod.string(),
  visitedAt: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListCountriesResponse = zod.array(ListCountriesResponseItem);

/**
 * @summary Add a visited country
 */
export const CreateCountryBody = zod.object({
  name: zod.string(),
  countryCode: zod.string(),
  visitedCities: zod.string(),
  reasonForVisit: zod.string(),
  travelCompanions: zod.string(),
  friendsFamilyMet: zod.string(),
  visitedAt: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
});

/**
 * @summary Get a visited country by id
 */
export const GetCountryParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCountryResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  countryCode: zod.string(),
  visitedCities: zod.string(),
  reasonForVisit: zod.string(),
  travelCompanions: zod.string(),
  friendsFamilyMet: zod.string(),
  visitedAt: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a visited country
 */
export const UpdateCountryParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCountryBody = zod.object({
  name: zod.string().optional(),
  countryCode: zod.string().optional(),
  visitedCities: zod.string().optional(),
  reasonForVisit: zod.string().optional(),
  travelCompanions: zod.string().optional(),
  friendsFamilyMet: zod.string().optional(),
  visitedAt: zod.string().optional(),
  latitude: zod.number().optional(),
  longitude: zod.number().optional(),
});

export const UpdateCountryResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  countryCode: zod.string(),
  visitedCities: zod.string(),
  reasonForVisit: zod.string(),
  travelCompanions: zod.string(),
  friendsFamilyMet: zod.string(),
  visitedAt: zod.string(),
  latitude: zod.number(),
  longitude: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a visited country
 */
export const DeleteCountryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get all map pins (posts with lat/lng) for the world map
 */
export const ListMapPinsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  excerpt: zod.string(),
  coverImageUrl: zod.string().nullable(),
  latitude: zod.number(),
  longitude: zod.number(),
  location: zod.string().nullable(),
  publishedAt: zod.string().nullable(),
});
export const ListMapPinsResponse = zod.array(ListMapPinsResponseItem);

/**
 * @summary Get travel stats summary
 */
export const GetStatsResponse = zod.object({
  totalCountries: zod.number(),
  totalPosts: zod.number(),
  totalCities: zod.number(),
  continents: zod.number(),
});