# lib/api-zod/src/generated/types/createCountryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Blog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCountryBody {
  name: string;
  countryCode: string;
  visitedCities: string;
  reasonForVisit: string;
  travelCompanions: string;
  friendsFamilyMet: string;
  visitedAt: string;
  latitude: number;
  longitude: number;
}