# lib/api-zod/src/generated/types/createPostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Blog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePostBody {
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  countryId?: number | null;
  /** @nullable */
  publishedAt?: string | null;
}