# lib/api-zod/src/generated/types/mapPin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Blog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MapPin {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  /** @nullable */
  coverImageUrl: string | null;
  latitude: number;
  longitude: number;
  /** @nullable */
  location: string | null;
  /** @nullable */
  publishedAt: string | null;
}