# lib/api-zod/src/generated/types/travelStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Blog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TravelStats {
  totalCountries: number;
  totalPosts: number;
  totalCities: number;
  continents: number;
}